+++ src/main/ipcs/index.ts
@@ -3,6 +3,7 @@ import './server/instance/createServerInstance';
 import './server/instance/createRemoteServerInstance';
 import './server/instance/testRemoteConnection';
 import './server/instance/editServerInstance';
+import './server/instance/editRemoteServerInstance';
 import './server/instance/deleteServerInstance';
 import './server/instance/duplicateServerInstance';
 import './server/instance/updateServerInstance';